% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal_cate_partially_linear.R
\name{fit_cate_hal_partially_linear}
\alias{fit_cate_hal_partially_linear}
\title{Doubly-robust nonparametric superefficient estimation of the partial conditional average treatment effect
using the highly adaptive lasso R-learner}
\usage{
fit_cate_hal_partially_linear(
  W,
  A,
  Y,
  weights = NULL,
  pi.hat = NULL,
  m.hat = NULL,
  formula_cate = NULL,
  max_degree_cate = 1,
  smoothness_orders_cate = 1,
  num_knots_cate = c(50),
  sl3_Lrnr_pi.hat = NULL,
  sl3_Lrnr_m.hat = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{W}{A numeric matrix of covariate values.}

\item{A}{A numeric vector of treatment values. Can be binary or continuous.}

\item{Y}{A numeric vector of outcome values.}

\item{weights}{(Optional) A numeric vector of observation weights.}

\item{pi.hat}{A numeric vector containing estimated propensity scores `pi(W) := P(A=1 | W)`.}

\item{m.hat}{A numeric vector containing estimates of treatment-marginalized outcome regression `m(W) := E[Y | W]`.}

\item{formula_cate}{(Optional) A `hal9001`-formatted formula object for the CATE to be passed to `formula_hal`.
By default, the CATE model is learned data-adaptively using MARS-based screening and HAL. See documentation for `fit_hal`.}

\item{max_degree_cate}{(Optional) Same as `max_degree` but for CATE model. See documentation for `fit_hal`.}

\item{smoothness_orders_cate}{(Optional) Same as `smoothness_orders` but for CATE model. See documentation for `fit_hal`.}

\item{num_knots_cate}{(Optional) Same as `num_knots` but for CATE model. See documentation for `fit_hal`.}

\item{sl3_Lrnr_pi.hat}{If `pi.hat` is not provided, a `Lrnr_base` object for estimation of `pi(W) := P(A=1 | W)`.}

\item{sl3_Lrnr_m.hat}{If `m.hat` is not provided, a `Lrnr_base` object for estimation of `m(W) := E[Y | W)`.}

\item{...}{Other arguments to be passed to `fit_hal`.}
}
\description{
This function estimates the Partial Conditional Average Treatment Effect (CATE) function `w -> tau(w) := E[Y | A=1, W=w] - E[Y | A=0, W=w]`
within the regression model `E[Y | A, W] = E[Y | A=0, W] + A * tau(W)`.
}
\details{
This method implements a doubly-robust and superefficient estimation technique for the Partial CATE using the highly adaptive lasso R-learner.
It can data-adaptivelly learn complex relationships in the CATE function, while benefiting from simpler structure and parsimony when present.
 By using the highly adaptive lasso, the method aims to provide robust and precise nonparametric inference of the Partial CATE.
}
