% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_cate_partially_linear.R
\name{fit_cate_lasso_partially_linear}
\alias{fit_cate_lasso_partially_linear}
\title{Doubly-robust nonparametric superefficient estimation of the conditional average treatment effect
using the lasso-based R-learner}
\usage{
fit_cate_lasso_partially_linear(
  W,
  A,
  Y,
  pi.hat = NULL,
  m.hat = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{W}{A numeric matrix of covariate values.}

\item{A}{A numeric vector of treatment values. Can be binary or continuous.}

\item{Y}{A numeric vector of outcome values.}

\item{pi.hat}{A numeric vector containing estimated propensity scores `pi(W) := P(A=1 | W)`.}

\item{m.hat}{A numeric vector containing estimates of treatment-marginalized outcome regression `m(W) := E[Y | W]`.}

\item{...}{Additional arguments to be passed to \code{\link[glmnet]{cv.glmnet}}.}
}
\description{
This function estimates the Conditional Average Treatment Effect (CATE) function `w -> tau(w) := E[Y | A=1, W=w] - E[Y | A=0, W=w]`
within the regression model `E[Y | A, W] = E[Y | A=0, W] + A * tau(W)`.
}
\details{
This method implements adaptive debiased machine learning of the Average Treatment Effect (ATE) through data-driven partially linear
model selection based on the LASSO (Least Absolute Shrinkage and Selection Operator). By incorporating the LASSO technique,
the method leverages learned structural information and promotes parsimony in the CATE function estimation. The approach
offers adaptivity and super-efficiency in nonparametric ATE inference.
}
