% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcate.R
\name{fit_hal_pcate}
\alias{fit_hal_pcate}
\title{Doubly-robust nonparametric superefficient estimation of the partial conditional average treatment effect
using the highly adaptive lasso R-learner
with adaptive MARS-based variable and interaction screening.}
\usage{
fit_hal_pcate(
  W,
  A,
  Y,
  Delta = NULL,
  weights = NULL,
  lrnr_A = NULL,
  lrnr_Y = NULL,
  max_degree = 3,
  max_degree_pi = max_degree,
  num_knots = c(sqrt(length(Y)), length(Y)^(1/3), length(Y)^(1/5)),
  smoothness_orders = 1,
  family_Y = c("auto", "gaussian", "binomial", "poisson"),
  family_A = c("auto", "gaussian", "binomial", "poisson"),
  formula_cate = NULL,
  max_degree_cate = max(max_degree - 1, 1),
  num_knots_cate = num_knots,
  smoothness_orders_cate = smoothness_orders,
  screen_variables = TRUE,
  screen_control = list(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{W}{A \code{matrix} of covariate values.}

\item{A}{A \code{numeric} vector of treatment values.
If binary then the CATE is estimated,
and otherwise the partial conditional average treatment effect is estimated.}

\item{Y}{A \code{numeric} vector of outcome values.}

\item{Delta}{(Not used)}

\item{weights}{(Optional) A \code{numeric} vector of observation weights.}

\item{max_degree}{For estimation of nuisance functions `E[Y|W]` and `E[X|W]`.
The maximum interaction degree of basis functions generated.
Passed to \code{\link[hal9001]{fit_hal}} function of \code{hal9001} package.}

\item{num_knots}{For estimation of nuisance functions `E[Y|W]` and `E[X|W]`.
Passed to \code{\link[hal9001]{fit_hal}} function of \code{hal9001} package.
A \code{numeric} vector of length \code{max_degree} where
the `d`-th entry specifies the number of univariable spline knot points to use
when generating the tensor-product basis functions of interaction degree `d`.}

\item{smoothness_orders}{For estimation of nuisance functions `E[Y|W]` and `E[X|W]`.
An integer taking values in (0,1,2,...)
specifying the smoothness order of the basis functions. See documentation for \code{\link[hal9001]{fit_hal}}.}

\item{family_Y}{A \code{\link[stats]{family}} object specifying the outcome type of the outcome \code{Y}.
This is passed internally to \code{\link[hal9001]{fit_hal}} when estimating `E[Y | W]`.}

\item{family_A}{A \code{\link[stats]{family}} object specifying the outcome type of the treatment \code{A}.
This is passed internally to \code{\link[hal9001]{fit_hal}} when estimating `E[A | W]`.}

\item{formula_cate}{(Optional) A \code{hal9001}-formatted \code{formula} object for the CATE/tau to be passed to \code{\link[hal9001]{formula_hal}}.
By default the CATE model is learned data-adaptivelly using MARS-based screening and HAL.}

\item{max_degree_cate}{(Optional) Same as \code{max_degree} but for CATE model.}

\item{num_knots_cate}{(Optional) Same as \code{num_knots} but for CATE model.}

\item{smoothness_orders_cate}{(Optional) Same as \code{smoothness_orders} but for CATE model.}

\item{screen_variables}{Highly recommended. See documentation for \code{\link[hal9001]{fit_hal}}.}

\item{...}{Other arguments to be passed to \code{\link[hal9001]{fit_hal}}.}

\item{screen_interactions}{Highly recommended. See documentation for \code{\link[hal9001]{fit_hal}}.}
}
\description{
This method estimates the conditional average treatment effect function `w - > tau(w)`
under the regression model `E[Y | A, W] = E[Y | A=0, W] + A * tau(W)`.
}
