% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_cate_partially_linear.R
\name{fit_lasso_cate_partially_linear}
\alias{fit_lasso_cate_partially_linear}
\title{Doubly-robust nonparametric superefficient estimation of the conditional average treatment effect
using the lasso-based R-learner}
\usage{
fit_lasso_cate_partially_linear(
  W,
  A,
  Y,
  pi.hat = NULL,
  m.hat = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{W}{A \code{numeric} \code{matrix} of covariate values.}

\item{A}{A \code{numeric} vector of treatment values. May be binary or continuous.}

\item{Y}{A \code{numeric} vector of outcome values.}

\item{pi.hat}{A \code{numeric} vector containing estimates of the propensity score `pi(W) := P(A=1 | W)`.}

\item{m.hat}{A \code{numeric} vector containing estimates of the treatment-marginalized outcome regression `m(W) := E[Y | W]`.}

\item{...}{Other arguments to be passed to \code{\link[glmnet]{cv.glmnet}}.}
}
\description{
This method estimates the conditional average treatment effect function `w - > tau(w) := E[Y | A=1, W =w] - E[Y | A=0, W = w]`
under the regression model `E[Y | A, W] = E[Y | A=0, W] + A * tau(W)`.
}
