% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_ate}
\alias{inference_ate}
\title{Estimates and confidence intervals for the ATE.}
\usage{
inference_ate(fit_cate, alpha = 0.05, return_cov_mat = FALSE)
}
\arguments{
\item{fit_cate}{A \code{hal_cate} object obtained from the function \code{fit_hal_cate}.}

\item{alpha}{Significant level for confidence intervals}

\item{return_cov_mat}{A \code{logical} for whether to return the asymptotic covariance matrix of the coefficient estimates.}
}
\description{
Estimates and confidence intervals for the ATE.
}
