% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_cate}
\alias{inference_cate}
\title{Estimates and confidence intervals for the projection of the CATE onto a user-specified parametric working model.}
\usage{
inference_cate(fit_cate, formula = ~1, alpha = 0.05, return_cov_mat = FALSE)
}
\arguments{
\item{fit_cate}{A \code{hal_cate} object obtained from the function \code{fit_hal_cate}.}

\item{formula}{A \code{formula} object specifying a working parametric model for the conditional average treatment effect.
For instance, `formula = ~ 1` specifies the marginal average treatment effect `E[CATE(W)]`.
More complex formula like `formula = ~ W1` specifies the best `W1`-linear approximation of the true CATE.}

\item{alpha}{Significant level for confidence intervals}

\item{return_cov_mat}{A \code{logical} for whether to return the asymptotic covariance matrix of the coefficient estimates.}
}
\description{
Estimates and confidence intervals for the projection of the CATE onto a user-specified parametric working model.
}
